requirements.txt
streamlit
pandas
PyGithub
Pillow
requests
rouge-score
bert-score
torch
transformers
aiohttp
tqdm
python-dotenv
pytest
